pyproject.toml
[project]
name = "solution-space-hacking"
version = "0.1.0"
description = "Solution Space Hacking project for exploring AI model solution space exploration during rollouts and training"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "pandas>=2.0.0",
    "aiohttp>=3.12.13",
    "datasets>=2.16.0",
    "huggingface_hub>=0.20.0",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "jsonlines>=4.0.0",
] 